+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Start-QualysScan, Get-QualysScans, Get-QualysScanSchedules, Add-QualysScanSchedule, Set-QualysAssetGroups
+
 ### Changed
 
 - Fixed incorrect comment-based help