+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
-## [1.2.0] - 2020-01-04
+## [1.2.0] - 2020-01-26
 
 ### Added